pyproject.toml
[project]
name = "anki-qb"
version = "0.1.0"
description = "Generate Anki flashcards from NAQT 'You Gotta Know' articles and Quiz Bowl questions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "lxml>=5.0.0",
    "more-itertools>=10.0.0",
    "google-genai>=1.0.0",
]
authors = [
    {name = "Adil"}
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
